+++ src/render.rs
@@ -8,6 +8,7 @@ use crate::TextInputStyle;
 use crate::edit::is_buffer_empty;
 use bevy::asset::AssetId;
 use bevy::asset::Assets;
+use bevy::camera::visibility::InheritedVisibility;
 use bevy::color::Alpha;
 use bevy::color::LinearRgba;
 use bevy::ecs::entity::Entity;
@@ -17,27 +18,26 @@ use bevy::ecs::system::Res;
 use bevy::ecs::system::ResMut;
 use bevy::image::TextureAtlasLayout;
 use bevy::input_focus::InputFocus;
-use bevy::math::Mat4;
+use bevy::math::Affine2;
 use bevy::math::Rect;
 use bevy::math::Vec2;
-use bevy::math::Vec3;
 use bevy::render::Extract;
 use bevy::render::sync_world::TemporaryRenderEntity;
-use bevy::render::view::InheritedVisibility;
 use bevy::sprite::BorderRect;
 use bevy::text::TextColor;
-use bevy::text::cosmic_text::Edit;
-use bevy::transform::components::GlobalTransform;
 use bevy::ui::CalculatedClip;
 use bevy::ui::ComputedNode;
-use bevy::ui::ComputedNodeTarget;
-use bevy::ui::ExtractedGlyph;
-use bevy::ui::ExtractedUiItem;
-use bevy::ui::ExtractedUiNode;
-use bevy::ui::ExtractedUiNodes;
-use bevy::ui::NodeType;
+use bevy::ui::ComputedUiTargetCamera;
 use bevy::ui::ResolvedBorderRadius;
-use bevy::ui::UiCameraMap;
+use bevy::ui::UiGlobalTransform;
+use bevy::ui_render::ExtractedGlyph;
+use bevy::ui_render::ExtractedUiItem;
+use bevy::ui_render::ExtractedUiNode;
+use bevy::ui_render::ExtractedUiNodes;
+use bevy::ui_render::NodeType;
+use bevy::ui_render::UiCameraMap;
+use bevy::ui_render::stack_z_offsets;
+use cosmic_text::Edit;
 
 pub fn extract_text_input_nodes(
     mut commands: Commands,
@@ -48,10 +48,10 @@ pub fn extract_text_input_nodes(
         Query<(
             Entity,
             &ComputedNode,
-            &GlobalTransform,
+            &UiGlobalTransform,
             &InheritedVisibility,
             Option<&CalculatedClip>,
-            &ComputedNodeTarget,
+            &ComputedUiTargetCamera,
             &TextInputLayoutInfo,
             &TextColor,
             &TextInputStyle,
@@ -99,12 +99,16 @@ pub fn extract_text_input_nodes(
             .editor
             .with_buffer(|buffer| Vec2::new(buffer.scroll().horizontal, 0.)); // buffer.scroll().vertical));
 
-        let transform = global_transform.affine()
-            * bevy::math::Affine3A::from_translation((-0.5 * uinode.size() - scroll).extend(0.));
+        let transform = Affine2::from(global_transform)
+            * Affine2::from_translation(uinode.size() * -0.5 - scroll);
 
         let node_rect = Rect::from_center_size(
-            global_transform.translation().truncate(),
-            uinode.size() * global_transform.scale().truncate(),
+            global_transform.translation,
+            uinode.size()
+                * Vec2::new(
+                    global_transform.matrix2.col(0).length(),
+                    global_transform.matrix2.col(1).length(),
+                ),
         );
 
         let clip = Some(
@@ -123,23 +127,23 @@ pub fn extract_text_input_nodes(
                 rect.size()
             } + 2. * Vec2::X;
             extracted_uinodes.uinodes.push(ExtractedUiNode {
-                stack_index: uinode.stack_index(),
-                color: LinearRgba::from(style.selection_color),
+                z_order: uinode.stack_index as f32 + stack_z_offsets::TEXT,
                 image: AssetId::default(),
                 clip,
                 extracted_camera_entity,
-                rect: Rect {
-                    min: Vec2::ZERO,
-                    max: size,
-                },
+                transform: transform * Affine2::from_translation(rect.center()),
                 item: ExtractedUiItem::Node {
+                    color: LinearRgba::from(style.selection_color),
                     atlas_scaling: None,
                     flip_x: false,
                     flip_y: false,
                     border_radius: ResolvedBorderRadius::ZERO,
                     border: BorderRect::ZERO,
                     node_type: NodeType::Rect,
-                    transform: transform * Mat4::from_translation(rect.center().extend(0.)),
+                    rect: Rect {
+                        min: Vec2::ZERO,
+                        max: size,
+                    },
                 },
                 main_entity: entity.into(),
                 render_entity: commands.spawn(TemporaryRenderEntity).id(),
@@ -180,27 +184,27 @@ pub fn extract_text_input_nodes(
             };
 
             let Some(rect) = texture_atlases
-                .get(&atlas_info.texture_atlas)
+                .get(atlas_info.texture_atlas)
                 .map(|atlas| atlas.textures[atlas_info.location.glyph_index].as_rect())
             else {
                 continue;
             };
 
             extracted_uinodes.glyphs.push(ExtractedGlyph {
-                transform: transform * Mat4::from_translation(position.extend(0.)),
+                color: color_out,
+                translation: *position,
                 rect,
             });
 
             extracted_uinodes.uinodes.push(ExtractedUiNode {
-                stack_index: uinode.stack_index(),
-                color: color_out,
-                image: atlas_info.texture.id(),
+                z_order: uinode.stack_index as f32 + stack_z_offsets::TEXT,
+                image: atlas_info.texture,
                 clip,
-                rect,
                 extracted_camera_entity,
                 item: ExtractedUiItem::Glyphs { range: start..end },
                 main_entity: entity.into(),
                 render_entity: commands.spawn(TemporaryRenderEntity).id(),
+                transform,
             });
 
             start = end;
@@ -217,28 +221,24 @@ pub fn extract_text_input_nodes(
             let width = style.cursor_width * scale_factor;
 
             extracted_uinodes.uinodes.push(ExtractedUiNode {
-                stack_index: uinode.stack_index(),
-                color,
+                z_order: uinode.stack_index as f32 + stack_z_offsets::TEXT,
                 image: AssetId::default(),
                 clip,
                 extracted_camera_entity,
-                rect: Rect {
-                    min: Vec2::ZERO,
-                    max: Vec2::new(width, cursor_height),
-                },
+                transform: transform
+                    * Affine2::from_translation(Vec2::new(x + 0.5 * width, y + 0.5 * line_height)),
                 item: ExtractedUiItem::Node {
+                    color,
                     atlas_scaling: None,
                     flip_x: false,
                     flip_y: false,
                     border_radius: ResolvedBorderRadius::ZERO,
                     border: BorderRect::ZERO,
                     node_type: NodeType::Rect,
-                    transform: transform
-                        * Mat4::from_translation(Vec3::new(
-                            x + 0.5 * width,
-                            y + 0.5 * line_height,
-                            0.,
-                        )),
+                    rect: Rect {
+                        min: Vec2::ZERO,
+                        max: Vec2::new(width, cursor_height),
+                    },
                 },
                 main_entity: entity.into(),
                 render_entity: commands.spawn(TemporaryRenderEntity).id(),
@@ -255,10 +255,10 @@ pub fn extract_text_input_prompts(
         Query<(
             Entity,
             &ComputedNode,
-            &GlobalTransform,
+            &UiGlobalTransform,
             &InheritedVisibility,
             Option<&CalculatedClip>,
-            &ComputedNodeTarget,
+            &ComputedUiTargetCamera,
             &TextInputPromptLayoutInfo,
             &TextColor,
             &TextInputBuffer,
@@ -301,12 +301,16 @@ pub fn extract_text_input_prompts(
 
         let color = prompt.color.unwrap_or(text_color.0).to_linear();
 
-        let transform = global_transform.affine()
-            * bevy::math::Affine3A::from_translation((-0.5 * uinode.size()).extend(0.));
+        let transform =
+            Affine2::from(global_transform) * Affine2::from_translation(-0.5 * uinode.size());
 
         let node_rect = Rect::from_center_size(
-            global_transform.translation().truncate(),
-            uinode.size() * global_transform.scale().truncate(),
+            global_transform.translation,
+            uinode.size()
+                * Vec2::new(
+                    global_transform.matrix2.col(0).length(),
+                    global_transform.matrix2.col(1).length(),
+                ),
         );
 
         let clip = Some(
@@ -321,20 +325,20 @@ pub fn extract_text_input_prompts(
         } in text_layout_info.glyphs.iter()
         {
             let rect = texture_atlases
-                .get(&atlas_info.texture_atlas)
+                .get(atlas_info.texture_atlas)
                 .unwrap()
                 .textures[atlas_info.location.glyph_index]
                 .as_rect();
             extracted_uinodes.glyphs.push(ExtractedGlyph {
-                transform: transform * Mat4::from_translation(position.extend(0.)),
+                color,
+                translation: *position,
                 rect,
             });
             extracted_uinodes.uinodes.push(ExtractedUiNode {
-                stack_index: uinode.stack_index(),
-                color,
-                image: atlas_info.texture.id(),
+                z_order: uinode.stack_index() as f32 + stack_z_offsets::TEXT,
+                transform,
+                image: atlas_info.texture,
                 clip,
-                rect,
                 item: ExtractedUiItem::Glyphs { range: start..end },
                 main_entity: entity.into(),
                 render_entity: commands.spawn(TemporaryRenderEntity).id(),